refactor(routes): move routes logic to own file
- add @types/url-search-params
- app component handles alias param on intial load
