+++ package.json
@@ -84,6 +84,7 @@
     "@types/node": "^12.11.1",
     "@types/react": "^16.9.9",
     "@types/react-dom": "^16.9.2",
+    "@types/url-search-params": "^1.1.0",
     "babel-eslint": "^10.0.2",
     "babel-jest": "^24.8.0",
     "babel-plugin-inline-react-svg": "^1.1.0",
